Merged PR 824602: Fix semaphore crash when acquiring the resources for the engine ram
In some dev builds, the engine ram's is more than the semaphore limit, causing a crash.

The deadlock will not happen if the engine uses all the semaphore value, we will be always running at least one pip even if the ram semaphore is full due to the engine.

Related work items: #2246664