Merged PR 827798: Allow skipping content in NestedDefinition
NestedDefinition is used to specify a nested layout on disk. When we try to flatten a nested definition, we process it in isolation, i.e., we ignore any items visited so far and the list of items visited during evaluation of nested definition will be discarded once it is processed. in simple cases, it works just fine. However, if the `contents` field contains conflicting entries we don't have an easy way to specify how this conflict should be resolved when one defines that nested definition. This PR adds a new field that allows users to specify list of items to exclude during flattening to resolve deployment collisions.