Merged PR 825342: Correct PathSet deserialization when sorting differs across platforms
Debug bits have a check that deserialized pathsets are sorted since they directly get loaded into a SortedReadOnlyArray. There are 2 problems with this:
1. When reloading on another computer, the sort order is not necessarily going to be the same. I hit this when loading a Linux created XLG on a Windows machine, but this can theoretically also happen more subtly on different computers running the same OS with different localization settings.
1. This check is not performed in release binaries. So in practice the check isn't really happening.

Assuming we want this scenario to work (reloading Linux created XLG files on Windows), we need to do something. The choice is to either correct the sort order when it is wrong or remove the sorting. I opted for the former since that seemed the more conservative approach.