Merged PR 830619: Attribute accesses shared opaque directory itself to a parent opaque
Accesses can happen to the shared opaque directory itself. If these are encoded into cache metadata, the consuming build blows up when trying to process them since it requires a relative path to the file under the opaque (and a direct access to the opaque would not yield a relative path)

This change skips attributing an access to the exact directory as the shared opaque. It will either land in a parent opaque or possibly generate a DFA if there is no parent opaque.

Related work items: #2253700