Merged PR 830941: [Cache GC] Download pages in parallel as we process them
Download pages in parallel as we process them.

This PR should roughly double the speed at which GC runs.

This is how long it takes to download a page, typically:

![image (2).png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/830941/attachments/image%20%282%29.png)

(i.e., somewhere along 7s to 10s)

This is how long it takes to process all 5k items in the page:

![image (3).png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/830941/attachments/image%20%283%29.png)

The code right now does download => process in a loop, so it roughly takes 20s to fetch 5k items and process them (and we do this across each shard). Moreover, worst case page download duration is around 20s, while worst case processing duration is around 1.5m, so basically we can download faster than we can process.

Since this PR moves the download to a separate task, we should -at least in principle- shave 10s off those 20s once we reach stability, as by the time we're ready to process we should already have a downloaded page in buffer.