Merged PR 832579: [EBPF]Add LRU cache on kernel side
Avoid sending repetitive events to user space so we alleviate the load on the ring buffer. We observed allocation problems when a big load of events are sent in a short period of time, and those typically involve the same operation and path (e.g. multiple write calls are used to write to the same file). This is similar to the event cache we use on user space, but some differences:
* It is a LRU cache with a fixed size. Events will get evicted after reaching the max size of the cache.
* The cache is local to each CPU. This makes eviction faster.