Merged PR 848773: Properly handle failure in adding output existence assertions from graph fragments
Adding output existence assertions from graph fragments can fail because of missing dependencies between graph fragments. For example, one can have the following spec for fragment 1:

```typescript
const outputDir = Unsafe.exOutputDirectory(outputDirPath);
const outputFileInDir = outputDir.assertExistence(p`someFile`);
```

where `outputDirPath` is produced by another fragment 2.

Unfortunately, the user does not specify fragment 2 as a dependency of fragment 1 when reading fragment 1, i.e.,

```typescript
const fragment1 = Transformer.readPipGraphFragment('fragment1');
const fragment2 = Transformer.readPipGraphFragment('fragment2');
```

In this PR, instead of having contract violation, we fail the graph construction with some error log messages.

Note that this PR doesn't solve the issue when the directory exists, but the fragment dependencies are not specified. For this bigger issue, not only for output existence assertions, but during the deserialization of input files/directories, we need to check (if they are not sources) that they are produced by the currently processed fragment or by its dependencies -- this for sure should be in a separate PR.