Merged PR 854394: Log telemetry for source rewrites with original content lost
We want to understand the cases where bxl misses the opportunity to see the content of an undeclared source before it gets rewritten. This is to assess the scope of the problem and decide our next move.

Collect input-related file operations (when undeclared source reads is on) whenever a source rewrite is present. The case we are interested in identifying is when a pip P rewrites a source file S. We want to know when P reads S before S is rewritten by P during its execution. If no other pip (ordered before the rewrite pip) read the source either, that means bxl didn't have the chance to observe the original content.