Merged PR 861752: Only replicate outputs to all workers in a distributed build
Office MetaBuild crashes in a single worker CloudBuild because `m_taskArrayPool` is null.

This is because `RemoteWorkerCount == 0` causes the role to be `DistributedBuildRoles.None` (in Engine.cs) which makes `IsDistributedOrchestrator` false which then causes `EnableDistribution` to not be called in the Engine which is what initializes `m_taskArrayPool`.

Adding a condition here should be correct as we simply don't need to replicate outputs to workers if there are no workers and subsequent logic triggers off of `MaterializeOutputsTasks` which is null as if it didn't happen.