Merged PR 857957: [EBPF] Ring buffer swapping
Encapsulate polling and sending events to the main queue into a RingBufferEntry. This entry exposes a callback to notify when the ringbuffer is about to get full. The handler is expected to create a new entry and deactivate the old one. This process can happen as many times as needed.