Merged PR 856985: Add blob daemon
This PR adds new daemon that can be used to upload build artifacts for azure storage. There are two observations that are important to this new daemon - 1) since all cache implementations used by buildxl  in one way or another are utilizing blob storage underneath, the build artifacts are stored in blob storage; and 2) by design,  by the time the upload pip is executed, its inputs (i.e., outputs of other pips) have been stored into cache. So, because of this, when asked to upload an artifact, the daemon can simply do a server side copy from cache blob storage to the target storage.

Cache auth is handled by bxl. To auth into target storage, a use must provide a name of env variable that contains a secret that allows blob uploads. The secret must be valid for the duration of a build.

Related work items: #2279863