Merged PR 863740: Put several daemon bins into a single shared folder
Put several daemon bins, i.e., Symbol, Materialization, and Blob, into a single shared folder.

Previously, each daemon had its own folder, consisting of daemon specs and a `bin/` folder for the executables.

The folder for each daemon is still there, but now, the executables are placed in the parent `Daemon.Bin` folder.

Validation:
* Selfhost: https://dev.azure.com/mseng/Domino/_build/results?buildId=30365677&view=logs&s=53185a19-ad5d-5172-ef74-5873532ca1a9
* Artifact publish: https://dev.azure.com/mseng/Domino/_build/results?buildId=30365677&view=logs&s=380ab8c6-4139-55ff-15c4-fb48260f4dca
* Selfhost distributed: https://dev.azure.com/mseng/Domino/_build/results?buildId=30366752&view=results

We currently do not apply this to Drop daemon due to a number of assembly indirection.

The resulting .nupkg for windows is now 411M