Merged PR 864588: [Detours] More robust Detours remote/brokered injection
Detours remote injection is flaky with various failure like

```
Failed to copy payload to process (error code: 0x0000012b)
```

or

```
Detours injection actually succeeded, but cannot open event 'Global\0000d2e4-0000142118e12bf6-S' to signal the caller: System.Threading.WaitHandleCannotBeOpenedException: No handle of the given name exists
```

For copy payload failure, it is well known that it is intermittent. Our sandbox executor and Detours services address this issue with retries. However, when it comes to remote/brokered injection, no retry is done. This is because the injection is performed by the managed `ProcessTreeContext` and that class only informs the native side with events indicating success or failure, i.e., there's no information about error code that would have allowed for retry when the error is partial copy.

To this end, this PR adds a retry mechanism in the managed `ProcessTreeContext`.

The other flakiness is caused by failure in opening events, created on the native side, by the managed `ProcessTreeContext`. These events are used to tell the native side if the injection performed by `ProcessTreeContext` is successful or not.

We added some information about the failure, and the error indicates that an event with the specified name doesn't exist. Some possibilities are:

1. the event with the specified name doesn't exists - which is weird.
2. the specified name is invalid - may be there's a weird character in the name
3. different security context.

Dotnet library treats 1 & 2 with `NameNotFound`. Thus, the PR adds more information by getting the last error code to understand whether it is 1 or 2.

For 3, this PR specifies explicitly security attributes when creating the events. These attributes are very permissive as they allow everyone to access. We may rethink about this.

Related work items: #2311363