Merged PR 864881: [Incremental Scheduling] Make graph-agnostic state consistent with non-agnostic dirty node tracker in presence of perpetually dirty pips
Previously, the graph-agnostic state can say that a pip is clean although it should not be because it should be dirtied by the presence of upstream perpetually dirty pips.

This incorrect information makes debugging challenging particularly when working on making the output existence assertions compatible with incremental scheduling.

Related work items: #2298997