Merged PR 868807: [EBPF] Canonicalize paths on kernel side
When computing a path from a dentry+mount, the outcome is always a canonicalized path. However, for the kernel functions we trace where a fs structure is not available and we deal with raw strings, paths may not be canonicalized since they are user input.
Canonicalize those paths on kernel side. This is about removing //, /., /.. and trailing slashes. With this change:
* Paths containing the aforementioned patterns get cleaned up and can effectively reach user side and become AbsolutePaths. Without this change, those accesses are lost
* Kernel side string caches have a higher chance of a hit
* Untracked scope membership can be done correctly. E.g. consider an untracked scope `/untracked/scope`. Without this change, `/untracked/./scope/nested` will be sent to user side anyway, and even worse, `/untracked/scope/../nested` won't be sent even though it is not untracked.