Merged PR 872169: [Detours] Avoid logging error for brokered injection error
We want brokered injections to be retriable at process creation level. To this end, we should avoid logging an error from the injector, i.e., `ProcessTreeContext`.

If we log an error DX, then this code in `Engine.cs` will call the scheduler to terminate immediately:

```csharp
m_trackingEventListener.RegisterInternalErrorAction(loggingContext, () => engineSchedule.Scheduler.TerminateForInternalError());
```

What that code does is trigger the scheduler cancellation token. That cancellation token is linked with the one we use for launching process in `SandboxedProcessPipExecutor.cs`; see:

```csharp
private async Task<SandboxedProcessPipExecutionResult> GetAndProcessResultAsync(
    ISandboxedProcess process,
    HashSet<AbsolutePath> allInputPathsUnderSharedOpaques,
    System.Diagnostics.Stopwatch sandboxPrepTime,
    CancellationToken cancellationToken)
{
    using var cancellationTokenSource = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken, m_context.CancellationToken);
    using var cancellationTokenRegistration = cancellationTokenSource.Token.Register(() => process.KillAsync());
```

That is, the process is killed immediately so there's no chance for doing a retry.

Related work items: #2311363