Merged PR 875594: [EBPF] Only use the string cache under ring buffer pressure
The ebpf string cache turned out to be very slow. In some scenarios where absent lookups/readlinks are very frequent (in the order of millions of accesses of that type per execution) pips can take double the time compared to interpose.
This PR makes it so we only use the string cache when the ring buffer has less than half of space, so the slowdown is justified.