Merged PR 878429: Catch Win32Exception potentially thrown by System.Diagnostics.Process.MainModule
Getting the filename from `process.MainModule.FileName` might throw this. The documentation says it happens if we try to call it on a 64 bit process from a 32 bit process, but this should never happen because we only run the main bxl process as a 64-bit process. It's unclear why this is happening, but we can catch it for now to avoid hitting this.

Related work items: #2326081