Merged PR 883418: Additional command line arguments for ADO builds via 1ES.Config + AdoBuildRunner
This PR adds capabilities for AdoBuildRunner to consume the BuildXL configuration directory that is pushed through 1ESPipelines.Config and exposed by 1ESPT through a special environment variable. With this, we can push configuration immediately to all builds that use the build tools installer. At the same time, we introduce one such configuration, `buildOverrides.json`, which for now will include additional command line arguments for both AdoBuildRunner and the build engine itself. Additional overrides could be added to this file or others in the same `buildxl/AdoBuildRunnerConfiguration` directory. The corresponding config-repo PR is !883422

Related work items: #2337362