Merged PR 886142: Cache build cache configuration via server mode
This is a performance optimization. Keep track of retrieved cache configurations in a static dictionary. This dictionary should survive the lifetime of a build when using the same instance of server mode.
Check that the cached tokens are fresh enough before using the cached value (at least 12hrs is the current threshold). Otherwise ping the cache endpoint again.